+++ .gitignore
@@ -24,3 +24,8 @@ hs_err_pid*
 /target
 /.classpath
 /.project
+*.iml
+.idea
+
+/*.iml
+/.idea
