+++ pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>org.gj.demo</groupId>
+	<groupId>com.jc.microservice</groupId>
 	<artifactId>gemfire-server</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<parent>
@@ -12,7 +12,6 @@
 
 	<properties>
 		<java.version>1.8</java.version>
-		<spring.version>5.0.1.RELEASE</spring.version>
 		<spring-data-releasetrain.version>Kay-SR1</spring-data-releasetrain.version>
 		<spring-shell.version>1.2.0.RELEASE</spring-shell.version>
 	</properties>
@@ -23,8 +22,13 @@
 			<artifactId>spring-boot-starter</artifactId>
 		</dependency>
 		<dependency>
+		<groupId>org.springframework.boot</groupId>
+		<artifactId>spring-boot-starter-web</artifactId>
+	    </dependency>
+		<dependency>
 			<groupId>org.springframework.data</groupId>
 			<artifactId>spring-data-gemfire</artifactId>
+			<version>2.0.2.RELEASE</version>
 		</dependency>
 		<dependency>
 			<groupId>org.projectlombok</groupId>
@@ -59,6 +63,11 @@
 			<version>2.9.3</version>
 		</dependency>
 		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-test</artifactId>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
 			<groupId>org.gj.demo</groupId>
 			<artifactId>gemfire-common</artifactId>
 			<version>0.0.1-SNAPSHOT</version>
@@ -66,11 +75,15 @@
 	</dependencies>
 	<repositories>
 		<repository>
-			<id>spring-releases</id>
-			<url>https://repo.spring.io/libs-release</url>
-		</repository>
-	</repositories>
-
+			<id>central</id >
+			<name>Maven Repository Switchboard</name >
+			<layout>default</layout >
+			<url>http://mvnrepository.com/</url >
+			<snapshots>
+				<enabled>false</enabled >
+			</snapshots >
+		</repository >
+	</repositories >
 	<build>
 		<plugins>
 			<plugin>
