+++ src/resources/application.yaml
@@ -1,0 +1,3 @@
+spring:
+ profiles:
+  active: dev
